Tiva is dumb and cant do floating-point...

+++ EvalBoards/TI/Tiva C Launchpad EK-TM4C123GXL/PlantMetrics/Ports.h
@@ -14,7 +14,7 @@ void Init_I2C0(void);
 uint32_t I2CReceive(uint32_t slave_addr, uint8_t reg);
 
 uint8_t Get_Temp(void); //Gets the soil temperature from the thermistor
-float Get_Brightness(void); //Gets the brightness from the VEML7700
+uint32_t Get_Brightness(void); //Gets the brightness from the VEML7700
 uint8_t Get_SoilMoisture(void);
 uint8_t Get_EnviromentInfo(char infoType);
 #endif
